backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
python-dotenv
importlib-metadata
reportlab
pytest
httpx
google-generativeai
